refactoring + handling kallsyms
